+++ tests/external_botocore/_mock_external_bedrock_server.py
