+++ tests/external_botocore/_test_bedrock_chat_completion.py
@@ -11,7 +11,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -36,7 +36,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "660d4de9-6804-460e-8556-4ab2a019d1e3",
                 "span_id": "span-id",
@@ -55,7 +55,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "660d4de9-6804-460e-8556-4ab2a019d1e3",
                 "span_id": "span-id",
@@ -76,7 +76,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -99,7 +99,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": "1234-0",
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "3bf1bb6b-b6f0-4901-85a1-2fa0e814440e",
                 "span_id": "span-id",
@@ -118,7 +118,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": "1234-1",
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "3bf1bb6b-b6f0-4901-85a1-2fa0e814440e",
                 "span_id": "span-id",
@@ -139,7 +139,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -161,7 +161,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "f354b9a7-9eac-4f50-a8d7-7d5d23566176",
                 "span_id": "span-id",
@@ -180,7 +180,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "f354b9a7-9eac-4f50-a8d7-7d5d23566176",
                 "span_id": "span-id",
@@ -201,7 +201,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionSummary"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "transaction_id": None,
                 "span_id": "span-id",
@@ -224,7 +224,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "c5188fb5-dc58-4cbe-948d-af173c69ce0d",
                 "span_id": "span-id",
@@ -243,7 +243,7 @@ chat_completion_expected_events = {
             {"type": "LlmChatCompletionMessage"},
             {
                 "id": None,  # UUID that varies with each run
-                "appName": "Python Agent Test (mlmodel_bedrock)",
+                "appName": "Python Agent Test (external_botocore)",
                 "conversation_id": "my-awesome-id",
                 "request_id": "c5188fb5-dc58-4cbe-948d-af173c69ce0d",
                 "span_id": "span-id",
