+++ tox.ini
@@ -97,7 +97,6 @@ envlist =
     redis-datastore_redis-{py37,py38,py39,py310,py311,pypy38}-redis{0400,latest},
     rediscluster-datastore_rediscluster-{py37,py311,pypy38}-redis{latest},
     python-datastore_sqlite-{py27,py37,py38,py39,py310,py311,pypy27,pypy38},
-    python-external_boto3-{py27,py37,py38,py39,py310,py311}-boto01,
     python-external_botocore-{py37,py38,py39,py310,py311}-botocorelatest,
     python-external_botocore-{py311}-botocore128,
     python-external_botocore-py310-botocore0125,
@@ -140,7 +139,6 @@ envlist =
     python-framework_starlette-{py37,py38}-starlette{002001},
     python-framework_starlette-{py37,py38,py39,py310,py311,pypy38}-starlettelatest,
     python-framework_strawberry-{py37,py38,py39,py310,py311}-strawberrylatest,
-    python-mlmodel_bedrock-{py37,py38,py39,py310,py311,pypy38},
     python-logger_logging-{py27,py37,py38,py39,py310,py311,pypy27,pypy38},
     python-logger_loguru-{py37,py38,py39,py310,py311,pypy38}-logurulatest,
     python-logger_loguru-py39-loguru{06,05},
@@ -252,10 +250,9 @@ deps =
     datastore_redis-redislatest: redis
     datastore_rediscluster-redislatest: redis
     datastore_redis-redis0400: redis<4.1
-    external_boto3-boto01: boto3<2.0
-    external_boto3-boto01: moto<2.0
-    external_boto3-py27: rsa<4.7.1
     external_botocore-botocorelatest: botocore
+    external_botocore-botocorelatest: boto3
+    external_botocore-botocorelatest: moto
     external_botocore-botocore128: botocore<1.29
     external_botocore-botocore0125: botocore<1.26
     external_botocore-{py37,py38,py39,py310,py311}: moto[awslambda,ec2,iam]<3.0
@@ -344,7 +341,6 @@ deps =
     framework_tornado: pycurl
     framework_tornado-tornadolatest: tornado
     framework_tornado-tornadomaster: https://github.com/tornadoweb/tornado/archive/master.zip
-    mlmodel_bedrock: boto3
     logger_loguru-logurulatest: loguru
     logger_loguru-loguru06: loguru<0.7
     logger_loguru-loguru05: loguru<0.6
@@ -439,7 +435,6 @@ changedir =
     datastore_redis: tests/datastore_redis
     datastore_rediscluster: tests/datastore_rediscluster
     datastore_sqlite: tests/datastore_sqlite
-    external_boto3: tests/external_boto3
     external_botocore: tests/external_botocore
     external_feedparser: tests/external_feedparser
     external_http: tests/external_http
@@ -464,7 +459,6 @@ changedir =
     framework_starlette: tests/framework_starlette
     framework_strawberry: tests/framework_strawberry
     framework_tornado: tests/framework_tornado
-    mlmodel_bedrock: tests/mlmodel_bedrock
     logger_logging: tests/logger_logging
     logger_loguru: tests/logger_loguru
     logger_structlog: tests/logger_structlog
